Create example file for unit tests
